backend and server connected
Frontend background and register page edited
