+++ frontend/swe-frontend/src/App.css
@@ -2,7 +2,13 @@
   text-align: center;
   background-color: rgb(39, 36, 36);
   background-image: url(musicWave.jpeg);
+  height: 100vh;
+  width: 100vw;
+}
 
+body {
+  height: 100vh;
+  width: 100vw;
 }
 
 .style{
