+++ frontend/swe-frontend/src/Reg.tsx
@@ -1,0 +1,43 @@
+import React, {SyntheticEvent, useState} from 'react'
+import { Connect, SendMessage } from "./api";
+
+const Reg = () => {
+    const [name, setName] = useState('');
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('');
+
+    const handleClick = () => { 
+        console.log("User loging in");
+        SendMessage("User loging in");
+        };
+
+        
+    const submit = (e: SyntheticEvent) => {
+        e.preventDefault();
+
+        console.log({
+            name,
+            email,
+            password
+        })
+    }
+    return (
+        <form onSubmit = {submit}>
+        <h1 className="text">Welcome to JAM. Sign in below</h1>
+        <input type="name" className="form-control" id="floatingInput" placeholder="First Last" required
+                        onChange = {e => setName(e.target.value)}/>
+                     
+
+          <input type="email" className="form-control" id="floatingInput" placeholder="name@example.com" required
+                onChange = {e => setEmail(e.target.value)}/>
+       
+          <input type="password" className="form-control" id="floatingPassword" placeholder="Password" required
+            onChange = {e => setPassword(e.target.value)}/>
+               
+        <button className="w-100 btn btn-lg btn-primary" type="submit" onClick={handleClick}>Sign in</button>
+        <p className="mt-5 mb-3 text-muted">&copy; 2023</p>
+      </form>
+    );
+};
+
+export default Reg;
